update add job page

+++ admin/.hta_slug/delete_job.php
@@ -1,12 +1,13 @@
 <?php
 require_once __DIR__ . '/../../.hta_slug/_header.php';
 require_once __DIR__ . '/../functions.php';
-// require_admin();
+require_admin();
 
 $id = (int)($_GET['id'] ?? 0);
 if ($id) {
     $stmt = $pdo->prepare("DELETE FROM jobs WHERE job_id = ?");
     $stmt->execute([$id]);
 }
-header('Location: dashboard.php');
+echo "<script>window.location.href='/admin/dashboard'</script>";
+// header('Location: dashboard.php');
 exit;
